Time: 15 ms (51.52%), Space: 38.1 MB (8.88%) - LeetHub

+++ 2116-check-if-a-parentheses-string-can-be-valid/2116-check-if-a-parentheses-string-can-be-valid.cpp
@@ -2,8 +2,9 @@ class Solution {
 public:
     bool canBeValid(string s, string locked) {
         if(s.size()%2!=0)return false;
-        vector<long int> v1;
-        vector<long int> v2;
+        vector<int> v1;
+        vector<int> v2;
+
         for(int i=0; i<s.size(); i++){
            if(locked[i]=='0')v1.push_back(i);
            else{
